+++ packages/eslint-plugin/configs/test-e2e.js
@@ -1,8 +1,6 @@
 const globals = require( 'globals' );
-const jestPlugin = require( 'eslint-plugin-jest' );
 
 module.exports = [
-	jestPlugin.configs[ 'flat/recommended' ],
 	{
 		languageOptions: {
 			globals: {