+++ packages/scripts/utils/index.js
@@ -9,7 +9,6 @@ const {
 	spawnScript,
 } = require( './cli' );
 const {
-	getJestOverrideConfigFile,
 	getVitestOverrideConfigFile,
 	getPhpFilePaths,
 	getProjectSourcePath,
@@ -17,7 +16,6 @@ const {
 	getWebpackEntryPoints,
 	hasBabelConfig,
 	hasCssnanoConfig,
-	hasJestConfig,
 	hasVitestConfig,
 	hasPostCSSConfig,
 	hasPrettierConfig,
@@ -42,7 +40,6 @@ module.exports = {
 	getArgFromCLI,
 	getArgsFromCLI,
 	getFileArgsFromCLI,
-	getJestOverrideConfigFile,
 	getVitestOverrideConfigFile,
 	getNodeArgsFromCLI,
 	getPackageProp,
@@ -56,7 +53,6 @@ module.exports = {
 	hasBabelConfig,
 	hasCssnanoConfig,
 	hasFileArgInCLI,
-	hasJestConfig,
 	hasVitestConfig,
 	hasPackageProp,
 	hasPostCSSConfig,